Update Denver.md
Vaccine supplies limited in Denver County and subject to the State of Colorado's phased distribution plan. **If you are eligible, you can sign up to reeive the vaccine at any provider, regardless of insurance.** If you are a current or recent patient of a Denver healthcare provider, and if you qualify, they may be reaching out to you proactively to schedule an appointment. 

**Some residents age 65-69, people age 16-64 with two or more high-risk conditions, pre-K-12 educators and child care providers, some state employees, and frontline essential workers may be eligible for the vaccine effective now (effective date 2/8/21). Visit [Colorado COVID-19 Vaccine Website](https://covid19.colorado.gov/vaccine) for details.**

* **Denver Health** visit [DenverHealth.org/CovidVaccine](https://www.denverhealth.org/covidvaccine)
* **Kaiser Permenente** visit [KP.org/CovidVaccine/Colorado](https://www.kp.org/covidvaccine/colorado)
* **SCL Health** visit [SCLHealth.org/COVIDVaccine](https://www.sclhealth.org/covidvaccine)
* **HealthOne** (Rose Medical Center) visit [HealthOneCares.com/Vaccine](https://www.healthonecares.com/vaccine)
* **UC Health** visit [UCHealth.org/covidvaccine](https://www.uchealth.org/covidvaccine)
* **Safeway** visit [Safeway.com/My-Vaccine-Communication.html](https://www.safeway.com/my-vaccine-communication.html) to sign up for COVID-19 vaccine information and notifications for when vaccine appointments become available at your store
* **King Soopers** visit [KingSoopers.com/i/coronavirus-update/vaccine](https://www.kingsoopers.com/i/coronavirus-update/vaccine) to see which pharmacies have received a limited supply of the COVID-19 vaccine. 

Colorado is currently in Phase 1 of the state's vaccination timeline. There are two vaccines being produced by Pfizer and Moderna. Both vaccines require two doses of the vaccine approximately 21 days apart. The companies report that both vaccines are around 95% effective. The city and county of Denver and the State of Colorado are ensuring that every person who receives their first dose will be able to access a second dose in the correct time frame. 

The vaccine will be free, but some providers may ask for insurance so they can bill for the vaccine administration fee. Medicare, Medicaid and private insurance will cover the cost of the COVID-19 vaccines. In addition, uninsured Coloradoans will ahve access to free vaccines. 

The vaccine will **not** be mandated by any agency or organization, as it is under the [Emergency Use Authorization](https://www.fda.gov/vaccines-blood-biologics/vaccines/emergency-use-authorization-vaccines-explained)